Skip attempting to register device when connecting with pre-existing ticketbooks (#2559)
As a workaround for the vpn-api currently returning an error when a
device attempts to re-register a device, skip registering the device as
part of connecting with pre-existing zknyms.

The completion of the device sync vpn-api call should trigger a register
device call if needed. It will be slower, but should be ok in the end.